backend/pyproject.toml
[tool.poetry]
name = "ai-rel-graph-backend"
version = "0.1.0"
description = "Backend for semantic graph visualization"
authors = ["Kaelem Deng <dengkaelem@gmail.com>", "Shawn Prather <contact@shawnprather.dev>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
redis = "^5.0.0"
celery = "^5.3.0"
sentence-transformers = "^2.2.2"
torch = "^2.1.0"
scikit-learn = "^1.3.0"
google-api-python-client = "^2.100.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
pypdf2 = "^3.0.0"
python-docx = "^1.0.0"
openpyxl = "^3.1.0"
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.0"
numpy = "^1.26.0"
pgvector = "^0.2.0"
kagglehub = "^0.3.0"
pycryptodome = "^3.20.0"
matplotlib = "^3.8.0"
networkx = "^3.1"
nltk = "^3.8.1"
keybert = "^0.8.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
email-validator = "^2.3.0"
httpx = "^0.25.0"
openai = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.10.0"
ruff = "^0.1.0"
mypy = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "redis.*",
    "sentence_transformers.*",
    "pgvector.*",
    "keybert.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"